Feature Selection
The following describes a process where data from the food_data dataset is prepared for machine learning modeling. First, unnecessary columns are removed using the drop() function. The LabelEncoder() function is then applied to convert the categorical target variable into numerical values, making it suitable for use in a machine learning model. The dataset is split into training and testing sets using train_test_split(). A RandomForestClassifier is then trained on the training set.

After training, the feature_importances_ attribute of the RandomForestClassifier is used to determine the importance of each feature. The sort_values() function sorts these features by their importance scores, and the top 10 features are selected.

The outcome of this process is a dataframe, top_10_features, that contains the names and importance scores of the ten most significant features, providing insights into which features are most critical for predicting the target variable in the model.