Gaussian Naive Bayes
This code trains a Gaussian Naive Bayes classifier on the selected features and calculates the training and testing accuracies. It then plots the training and testing accuracies to compare them.

- The GaussianNB() function from sklearn is used to train a Gaussian Naive Bayes classifier on the training data with selected features.

- The predict() function is used to make predictions for the training and testing sets.

- The accuracy_score() function is used to calculate the training and testing accuracies.

- The bar() function from matplotlib is used to plot the training and testing accuracies. The title(), ylabel(), and ylim() functions are used to set the title, y-axis label, and y-axis limits of the plot, respectively.

The resulting plot shows the training and testing accuracies for the Gaussian Naive Bayes model, which can be used to evaluate the performance of the model on the selected features.
The training accuracy is 0.811 and the testing accuracy is 0.808, which indicates that the model is performing well on both the training and testing sets.