Dropping Useless Column (User_ID) from the Dataset
This code drops the 'User_ID' column from the merged_data DataFrame using the drop() function. The axis=1 parameter specifies that the column should be dropped (as opposed to a row, which would be axis=0).

The resulting DataFrame, merged_data_dropped, is then displayed using the head() function, which shows the first few rows of the DataFrame.