Linear Regression
This code prepares the data for a linear regression model by encoding the 'Gender' column as 1 for 'male' and 0 for 'female', and splitting the data into features (X) and target (y). The data is then split into training and testing sets using the train_test_split() function.
A linear regression model is created using the LinearRegression() function and trained on the training data using the fit() function. Predictions are then made on the training and testing data using the predict() function.
The training and testing accuracy is calculated using the r2_score() function, and a plot is created to compare the predicted and actual values for the training and testing data.