Support Vector Regression
This code creates a support vector regression model using the SVR() function and trains it on the training data using the fit() function. Predictions are then made on the training and testing data using the predict() function.

The training and testing accuracy is calculated using the r2_score() function, and a plot is created to compare the predicted and actual values for the training and testing data.

Finally, the training and testing accuracy is printed using the print() function.