Final Model
This code defines a function predict_calories() that takes a trained machine learning model as input and uses it to predict the number of calories burned based on user inputs.

The function first takes user inputs for gender, age, height, weight, duration, heart rate, and body temperature. The gender input is then encoded as 1 for 'male' and 0 for 'female'.

A numpy array of inputs is then created, and the predict() function of the input model is used to predict the number of calories burned based on these inputs.

Finally, the predicted calories are printed using the print() function.

The function is then called with the existing trained model rf_model to make a prediction for a specific set of user inputs.